refactor(app): Reestrutura projeto, implementa concorrência e melhora robustez
