+++ app/database.py
@@ -1,43 +1,96 @@
 import os
 from datetime import datetime
-from pymongo import MongoClient
-from fastapi import HTTPException
+from motor.motor_asyncio import AsyncIOMotorClient
+from fastapi import HTTPException, Depends
 
 MONGO_URI = os.getenv("MONGO_URI")
 
 if not MONGO_URI:
-    print("❌ ERRO: A variável de ambiente MONGO_URI não foi definida.")
-    exit()
+    raise HTTPException(status_code=500, detail="Internal Server Error: A variável de ambiente MONGO_URI não foi definida.")
 
 try:
-    sync_client = MongoClient(MONGO_URI)
-    sync_db = sync_client.techmatch_logs
-    sync_log_collection = sync_db.requests
+    async_client = AsyncIOMotorClient(MONGO_URI)
+    async_db = async_client.techmatch_logs
+    async_log_collection = async_db.requests
     
-    print("✅ Conexão com o MongoDB (Motor e PyMongo) pronta para ser usada.")
 except Exception as e:
-    print(f"❌ Falha na configuração do cliente MongoDB: {e}")
     raise HTTPException(status_code=500, detail="Internal Server Error: Erro ao acessar banco de dados. Tente novamente mais tarde.")
 
-def log_request_sync(log_data: dict):
+
+async def check_database_connection() -> bool:
+    """
+    Verifica se a conexão com o banco de dados está funcionando.
+    
+    Returns:
+        bool: True se a conexão estiver OK, False caso contrário
+        
+    Raises:
+        HTTPException: Se houver erro crítico na conexão
+    """
+    try:
+        # Tenta fazer um ping no servidor MongoDB
+        await async_client.admin.command('ping')
+        
+        # Verifica se o banco de dados existe e está acessível
+        db_names = await async_client.list_database_names()
+        
+        # Verifica se a coleção existe ou pode ser criada
+        collections = await async_db.list_collection_names()
+        
+        return True
+        
+    except Exception as e:
+        raise HTTPException(
+            status_code=503, 
+            detail="Service Unavailable: Banco de dados indisponível. Tente novamente mais tarde."
+        )
+
+
+async def get_database_dependency():
+    """
+    Dependência do FastAPI para verificar a conexão com o banco de dados.
+    
+    Esta função será usada como dependência nas rotas que precisam do banco.
+    
+    Returns:
+        bool: True se a conexão estiver OK
+        
+    Raises:
+        HTTPException: Se o banco estiver indisponível
+    """
+    return await check_database_connection()
+
+
+async def log_request_async(log_data: dict):
     """
     Insere um documento de log na coleção 'requests'.
     """
     try:
         log_data["timestamp"] = datetime.now()
-        sync_log_collection.insert_one(log_data)
+        await async_log_collection.insert_one(log_data)
         print(f"-> Log da requisição '{log_data.get('request_id')}' salvo com sucesso.")
     except Exception as e:
-        print(f"❌ ERRO ao salvar log síncrono no MongoDB: {e}")
         raise HTTPException(status_code=500, detail="Internal Server Error: Erro ao acessar banco de dados. Tente novamente mais tarde.")
     
-def get_analysis_by_request_id_sync(request_id: str) -> dict | None:
+
+async def get_analysis_by_request_id_async(request_id: str) -> dict | None:
     """
     Busca um log de requisição pelo seu request_id.
     """
     try:
-        result = sync_log_collection.find_one({"request_id": request_id}, {"_id": 0})
+        result = await async_log_collection.find_one({"request_id": request_id}, {"_id": 0})
         return result
     except Exception as e:
-        print(f"❌ ERRO ao buscar log síncrono no MongoDB: {e}")
         raise HTTPException(status_code=500, detail="Internal Server Error: Erro ao acessar banco de dados. Tente novamente mais tarde.")
+
+
+async def close_database_connection():
+    """
+    Fecha a conexão com o banco de dados.
+    
+    Esta função deve ser chamada no shutdown da aplicação.
+    """
+    try:
+        async_client.close()
+    except Exception as e:
+        raise HTTPException(status_code=500, detail="Internal Server Error: Erro ao fechar conexão com MongoDB. Tente novamente mais tarde.")
